django/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-environ==0.11.2
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==26.1.0
Markdown==3.6
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.1
